backend/requirement.txt
fastapi
uvicorn[standard]
mlflow
prometheus_client
psutil
apscheduler
pillow
